[MM-42906] - License receives invalid date in the console when license expiration check runs
