+++ components/admin_console/license_settings/license_settings.tsx
@@ -5,7 +5,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable max-lines */
 import React from 'react';
-import {FormattedDate, FormattedTime} from 'react-intl';
 
 import {ClientConfig, ClientLicense} from 'mattermost-redux/types/config';
 import {ActionResult} from 'mattermost-redux/types/actions';
@@ -282,16 +281,6 @@ export default class LicenseSettings extends React.PureComponent<Props, State> {
     render() {
         const {license, upgradedFromTE, isDisabled} = this.props;
 
-        const issued = (
-            <>
-                <FormattedDate value={new Date(parseInt(license.IssuedAt, 10))}/>
-                {' '}
-                <FormattedTime value={new Date(parseInt(license.IssuedAt, 10))}/>
-            </>
-        );
-        const startsAt = <FormattedDate value={new Date(parseInt(license.StartsAt, 10))}/>;
-        const expiresAt = <FormattedDate value={new Date(parseInt(license.ExpiresAt, 10))}/>;
-
         let leftPanel = null;
         let rightPanel = null;
 
@@ -324,9 +313,6 @@ export default class LicenseSettings extends React.PureComponent<Props, State> {
                     upgradedFromTE={upgradedFromTE}
                     license={license}
                     isTrialLicense={isTrialLicense(license)}
-                    issued={issued}
-                    startsAt={startsAt}
-                    expiresAt={expiresAt}
                     handleRemove={this.confirmLicenseRemoval}
                     isDisabled={isDisabled}
                     removing={this.state.removing}
